requirements.txt
apipkg==1.5
APScheduler==3.8.1
asgiref==3.3.4
attrs==21.2.0
backports.zoneinfo==0.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coloredlogs==15.0.1
convertdate==2.3.2
coverage==5.5
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.11
django-apscheduler==0.6.0
django-sendgrid-v5==0.9.1
execnet==1.8.0
factory-boy==2.12.0
Faker==8.2.1
future==0.18.2
hijri-converter==2.1.3
holidays==0.11.2
humanfriendly==10.0
idna==2.10
iniconfig==1.1.1
korean-lunar-calendar==0.2.1
oauthlib==3.1.0
ordered-set==4.0.2
packaging==20.9
pluggy==0.13.1
psycopg2==2.9.1
py==1.10.0
pycparser==2.20
PyJWT==2.0.1
PyMeeus==0.5.11
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.3.0
pytest-forked==1.3.0
pytest-xdist==2.2.1
python-dateutil==2.8.1
python-http-client==3.3.2
python3-openid==3.2.0
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
requests==2.25.1
requests-oauthlib==1.3.0
sendgrid==6.7.0
six==1.15.0
slack-sdk==3.14.1
social-auth-core==4.1.0
sqlparse==0.4.2
starkbank-ecdsa==1.1.0
text-unidecode==1.3
toml==0.10.2
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
